+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/updates/queues/PendingUpdateType.java
@@ -25,4 +25,5 @@ public enum PendingUpdateType {
     MinBlockTime,
     RootKeys,
     FinalizationCommitteeParameters,
+    ValidatorScoreParameters,
 }